requirements.txt
pydantic
beautifulsoup4
requests
neo4j
pymongo
python-dotenv
fastapi
uvicorn
pandas